backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
python-dotenv
cryptography
pandas
joblib
scikit-learn==1.6.1
google-cloud-vision
python-multipart
rapidfuzz
numpy